# src/generated/openapi/models.py
# generated by datamodel-codegen:
#   filename:  openapi.yml
#   timestamp: 2026-04-10T09:27:07+00:00

from __future__ import annotations

from uuid import UUID

from pydantic import AnyUrl, AwareDatetime, BaseModel, Field


class Book(BaseModel):
    title: str = Field(..., examples=['Spring Boot in Action'], max_length=255)
    author: str = Field(..., examples=['Craig Walls'], max_length=255)
    isbn: str = Field(..., examples=[9781617292545], max_length=20)
    publication_year: int | None = Field(None, alias='publicationYear', examples=[2016])
    publisher: str | None = Field(
        None, examples=['Manning Publications'], max_length=255
    )
    price: float | None = Field(None, examples=[39.99])
    quantity: int | None = Field(None, examples=[25], ge=0)


class BookResponse(BaseModel):
    id: UUID | None = Field(None, examples=['cfa5fde1-ad43-4463-9b14-8580f3fe04a8'])
    title: str | None = Field(None, examples=['Spring Boot in Action'])
    author: str | None = Field(None, examples=['Craig Walls'])
    isbn: str | None = Field(None, examples=[9781617292545])
    publication_year: int | None = Field(None, alias='publicationYear', examples=[2016])
    publisher: str | None = Field(None, examples=['Manning Publications'])
    price: float | None = Field(None, examples=[39.99])
    quantity: int | None = Field(None, examples=[25])
    created_at: AwareDatetime | None = Field(None, alias='createdAt')
    updated_at: AwareDatetime | None = Field(None, alias='updatedAt')


class Sort(BaseModel):
    sorted: bool | None = Field(None, examples=[True])
    unsorted: bool | None = Field(None, examples=[False])
    empty: bool | None = Field(None, examples=[False])


class Pageable(BaseModel):
    page_number: int | None = Field(None, alias='pageNumber', examples=[0])
    page_size: int | None = Field(None, alias='pageSize', examples=[20])
    sort: Sort | None = None
    offset: int | None = Field(None, examples=[0])
    paged: bool | None = Field(None, examples=[True])
    unpaged: bool | None = Field(None, examples=[False])


class PagedBookResponse(BaseModel):
    content: list[BookResponse] | None = None
    pageable: Pageable | None = None
    total_pages: int | None = Field(None, alias='totalPages', examples=[5])
    total_elements: int | None = Field(None, alias='totalElements', examples=[100])
    last: bool | None = Field(None, examples=[False])
    size: int | None = Field(None, examples=[20])
    number: int | None = Field(None, examples=[0])
    sort: Sort | None = None
    number_of_elements: int | None = Field(
        None, alias='numberOfElements', examples=[20]
    )
    first: bool | None = Field(None, examples=[True])
    empty: bool | None = Field(None, examples=[False])


class Problem(BaseModel):
    type: AnyUrl | None = Field(
        'about:blank', examples=['https://api.example.com/errors/validation-error']
    )
    """
    A URI reference that identifies the problem type
    """
    title: str | None = Field(None, examples=['Validation Error'])
    """
    A short, human-readable summary of the problem type
    """
    status: int | None = Field(None, examples=[400])
    """
    The HTTP status code
    """
    detail: str | None = Field(None, examples=['The ISBN provided is not valid'])
    """
    A human-readable explanation specific to this occurrence of the problem
    """
    instance: AnyUrl | None = Field(None, examples=['/api/books/isbn/invalid-isbn'])
    """
    A URI reference that identifies the specific occurrence of the problem
    """