# src/resources/swagger/openapi.yml
openapi: 3.0.3
info:
  title: Bookstore API
  description: REST API for Bookstore Application
  version: "1.0.0" # This will be replaced with the actual version during build
  contact:
    name: Developer Team
    email: developer@example.com
  license:
    name: Proprietary
servers:
  - url: "/api"
    description: API base path
tags:
  - name: books
    description: Book management operations
paths:
  /books:
    get:
      summary: List all books
      description: Returns a list of all books in the Bookstore
      operationId: getAllBooks
      tags:
        - books
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponse"
    post:
      summary: Add a new book
      description: Adds a new book to the Bookstore
      operationId: createBook
      tags:
        - books
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "400":
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /books/{bookId}:
    parameters:
      - name: bookId
        in: path
        description: ID of book to use
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get book by ID
      description: Returns a single book
      operationId: getBookById
      tags:
        - books
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "404":
          description: Book not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    put:
      summary: Update book
      description: Update an existing book
      operationId: updateBook
      tags:
        - books
      requestBody:
        description: Updated book object
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "400":
          description: Invalid input
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Book not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      summary: Delete book
      description: Delete a book from the Bookstore
      operationId: deleteBook
      tags:
        - books
      responses:
        "204":
          description: Successful operation
        "404":
          description: Book not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /books/isbn/{isbn}:
    parameters:
      - name: isbn
        in: path
        description: ISBN of book to find
        required: true
        schema:
          type: string
    get:
      summary: Find book by ISBN
      description: Returns a single book
      operationId: getBookByIsbn
      tags:
        - books
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookResponse"
        "404":
          description: Book not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /books/author/{author}:
    parameters:
      - name: author
        in: path
        description: Author of books to find
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: Page number (zero-based)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: Number of records per page
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      - name: sort
        in: query
        description: Sort field and direction (e.g. title,asc or price,desc)
        required: false
        schema:
          type: string
          default: "title,asc"
    get:
      summary: Find books by author
      description: Returns all books by a specific author, or an empty list if none found
      operationId: getBooksByAuthor
      tags:
        - books
      responses:
        "200":
          description: Successful operation (may be an empty list if no books found)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedBookResponse"
components:
  schemas:
    Book:
      type: object
      required:
        - title
        - author
        - isbn
      properties:
        title:
          type: string
          example: Spring Boot in Action
          maxLength: 255
        author:
          type: string
          example: Craig Walls
          maxLength: 255
        isbn:
          type: string
          example: 9781617292545
          maxLength: 20
        publicationYear:
          type: integer
          format: int32
          example: 2016
        publisher:
          type: string
          example: Manning Publications
          maxLength: 255
        price:
          type: number
          format: double
          example: 39.99
        quantity:
          type: integer
          format: int32
          minimum: 0
          example: 25
    BookResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: cfa5fde1-ad43-4463-9b14-8580f3fe04a8
        title:
          type: string
          example: Spring Boot in Action
        author:
          type: string
          example: Craig Walls
        isbn:
          type: string
          example: 9781617292545
        publicationYear:
          type: integer
          format: int32
          example: 2016
        publisher:
          type: string
          example: Manning Publications
        price:
          type: number
          format: double
          example: 39.99
        quantity:
          type: integer
          format: int32
          example: 25
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PagedBookResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/BookResponse"
        pageable:
          type: object
          properties:
            pageNumber:
              type: integer
              example: 0
            pageSize:
              type: integer
              example: 20
            sort:
              type: object
              properties:
                sorted:
                  type: boolean
                  example: true
                unsorted:
                  type: boolean
                  example: false
                empty:
                  type: boolean
                  example: false
            offset:
              type: integer
              example: 0
            paged:
              type: boolean
              example: true
            unpaged:
              type: boolean
              example: false
        totalPages:
          type: integer
          example: 5
        totalElements:
          type: integer
          example: 100
        last:
          type: boolean
          example: false
        size:
          type: integer
          example: 20
        number:
          type: integer
          example: 0
        sort:
          type: object
          properties:
            sorted:
              type: boolean
              example: true
            unsorted:
              type: boolean
              example: false
            empty:
              type: boolean
              example: false
        numberOfElements:
          type: integer
          example: 20
        first:
          type: boolean
          example: true
        empty:
          type: boolean
          example: false
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          default: about:blank
          example: https://api.example.com/errors/validation-error
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: Validation Error
        status:
          type: integer
          description: The HTTP status code
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
          example: The ISBN provided is not valid
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem
          example: /api/books/isbn/invalid-isbn